feat: prepare resolver and interpreter for resolving

+++ src/ast.rs
@@ -282,7 +282,7 @@ pub struct Interpreter<'a, W: std::io::Write> {
     environment: Rc<RefCell<Environment>>,
     callables: Box<Catalogue<'a>>,
     writer: W,
-    locals: HashMap<&'a str, i32>,
+    locals: HashMap<&'a str, usize>,
 }
 
 impl<'a, W: std::io::Write> Interpreter<'a, W> {
@@ -316,7 +316,7 @@ impl<'a, W: std::io::Write> Interpreter<'a, W> {
         })
     }
 
-    pub fn resolve(&mut self, expr: &Expr<'a>, depth: i32) {
+    pub fn resolve(&mut self, expr: &Expr<'a>, depth: usize) {
         match &expr.kind {
             ExprKind::Call(Token::Identifier(name), _, _) => {
                 self.locals.insert(name, depth);
